requirements.txt
ultralytics==8.0.149
streamlit==1.25.0
Pillow==10.0.0